vqvae/requirements.txt
certifi==2019.3.9
cffi==1.12.3
# mkl-fft==1.0.12
# mkl-random==1.0.2
numpy==1.16.4
olefile==0.46
Pillow==6.0.0
pycparser==2.19
six==1.12.0
# torch==1.1.0
torchvision==0.2.1
